refactor: 우승자 출력 수정
