+++ src/main/java/racingcar/model/RandomGenerator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.model;
 
 import java.util.Random;
