+++ src/test/java/racingcar/model/RandomGeneratorTest.java
@@ -1,6 +1,7 @@
+package racingcar.model;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.RandomGenerator;
 
 import static org.assertj.core.api.Assertions.assertThat;
